requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
python-dotenv==1.0.1
alembic==1.13.1
pytest==8.1.1
httpx==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3
apscheduler==3.10.4
spacy==3.7.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.3.4 